+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## unreleased
 
 ### Added
+
+- Ability to create new info file